+++ frontend/index.js
@@ -1,5 +1,6 @@
-import { CarShop, CarsList, CarCard } from './packages/car-shop';
+import { CarShop, CarsList, CarCard, CarDetails } from './packages/car-shop';
 
 customElements.define('car-shop', CarShop);
 customElements.define('cars-list', CarsList);
 customElements.define('car-card', CarCard);
+customElements.define('car-details', CarDetails);