+++ frontend/packages/car-shop/__element-definitions/car-details.js
@@ -1,0 +1,3 @@
+import { CarDetails } from '../src/CarDetails.js';
+
+window.customElements.define('car-details', CarDetails);