+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
-**1.0**
+**v1.0**
 - feat: List all cars across all warehouses
+
+**v2.0**
+- feat: Show details of car on click of 'Show More
+- feat: Accessiblity improvement for the website